(BSR)[API] refactor: CORE/USER/API SPLIT INTO MULTIPLE FILES
