+++ api/src/pcapi/core/users/api/__init__.py
@@ -1,0 +1,2 @@
+from .api import *
+from .suspension import *
